+++ src/features/email/services/email-data.ts
@@ -1,4 +1,4 @@
-import { TEmail } from '../types/email';
+import { TEmail } from '../types/email.types';
 
 export const emailData: TEmail[] = [
   {
@@ -7,7 +7,7 @@ export const emailData: TEmail[] = [
     subject: 'Meeting Tomorrow',
     preview:
       "Hi, Let's have a meeting tomorrow to discuss the project. I've been reviewing the project details and have some ideas I'd like to share. It's crucial that we align on our next steps to ensure the project's success. Please come prepared with any questions or insights you may have. Looking forward to our meeting! Best regards, Adrian",
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: false,
     isStarred: true,
     isImportant: true,
@@ -19,7 +19,7 @@ export const emailData: TEmail[] = [
     sender: 'Ethan Gold',
     subject: 'Meeting Rescheduled',
     preview: 'Hey, The meeting originally scheduled for 2 PM has been moved to 3:30 PM. Let m...',
-    date: `${new Date()}`,
+    date: `2025-03-27T09:31:25.000Z`,
     isRead: false,
     isStarred: false,
     isImportant: false,
@@ -31,7 +31,7 @@ export const emailData: TEmail[] = [
     subject: 'Reminder: Submit Your Report',
     preview:
       'Hello, Just a friendly reminder to submit your report by EOD today. Let us know if...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: false,
     isStarred: true,
     isImportant: false,
@@ -42,7 +42,7 @@ export const emailData: TEmail[] = [
     sender: 'Luca Fischer',
     subject: 'Subject: Lunch Plans?',
     preview: 'Hey, Are we still on for lunch tomorrow? Let me know what time works for you!...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: true,
@@ -53,7 +53,7 @@ export const emailData: TEmail[] = [
     sender: 'Julian Schmid',
     subject: 'Quick Question',
     preview: "Hey, Do you have the latest version of the project doc? I can't seem to find it...",
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -64,7 +64,7 @@ export const emailData: TEmail[] = [
     sender: 'Lena Weber',
     subject: 'Project Update',
     preview: 'Hi, Here’s the latest update on the project. We’ve made significant progress...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -75,7 +75,7 @@ export const emailData: TEmail[] = [
     sender: 'Noah Becker',
     subject: 'Feedback Needed',
     preview: 'Hello, Could you please review the attached document and provide your feedback...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -86,7 +86,7 @@ export const emailData: TEmail[] = [
     sender: 'Mia Hoffmann',
     subject: 'Team Meeting Agenda',
     preview: 'Hey, Here’s the agenda for our team meeting tomorrow. Please review it before...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: true,
     isImportant: false,
@@ -97,7 +97,7 @@ export const emailData: TEmail[] = [
     sender: 'Elias Richter',
     subject: 'New Feature Announcement',
     preview: 'Hi Team, We’re excited to announce the launch of a new feature. Check it out...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: true,
@@ -108,7 +108,7 @@ export const emailData: TEmail[] = [
     sender: 'Clara Schulz',
     subject: 'Holiday Party Details',
     preview: 'Hello, Here are the details for our upcoming holiday party. Looking forward to...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -119,7 +119,7 @@ export const emailData: TEmail[] = [
     sender: 'David Wolf',
     subject: 'Client Meeting Notes',
     preview: 'Hi, Attached are the notes from our client meeting. Please review and let me...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -130,7 +130,7 @@ export const emailData: TEmail[] = [
     sender: 'Anna Braun',
     subject: 'Project Deadline Reminder',
     preview: 'Hello, Just a reminder that the project deadline is approaching. Please ensure...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -141,7 +141,7 @@ export const emailData: TEmail[] = [
     sender: 'Paul König',
     subject: 'Training Session Feedback',
     preview: 'Hey, Thank you for attending the training session. Your feedback is appreciated...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: true,
     isImportant: false,
@@ -152,7 +152,7 @@ export const emailData: TEmail[] = [
     sender: 'Laura Gross',
     subject: 'Team Outing Plans',
     preview: 'Hi, We’re planning a team outing next month. Any suggestions on where to go?...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: true,
@@ -163,7 +163,7 @@ export const emailData: TEmail[] = [
     sender: 'Simon Jung',
     subject: 'New Policy Announcement',
     preview: 'Hello, We’re introducing a new company policy effective next week. Please read...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -174,7 +174,7 @@ export const emailData: TEmail[] = [
     sender: 'Hannah Keller',
     subject: 'Sales Report',
     preview: 'Hi, Attached is the monthly sales report. Please review it and let me know if...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -185,7 +185,7 @@ export const emailData: TEmail[] = [
     sender: 'Maximilian Vogel',
     subject: 'Project Proposal Review',
     preview: 'Hello, Could you please review the attached project proposal and provide your...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -196,7 +196,7 @@ export const emailData: TEmail[] = [
     sender: 'Lea Hartmann',
     subject: 'Website Maintenance',
     preview: 'Hey, We’ll be performing website maintenance this weekend. Please be aware of...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: true,
     isImportant: false,
@@ -207,7 +207,7 @@ export const emailData: TEmail[] = [
     sender: 'Felix Lange',
     subject: 'Customer Feedback',
     preview: 'Hi, We’ve received some positive customer feedback. Check out the attached...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: true,
@@ -218,7 +218,7 @@ export const emailData: TEmail[] = [
     sender: 'Emilia Wagner',
     subject: 'Office Closure Notice',
     preview: 'Hello, Please be informed that the office will be closed on [Date] for [Reason]...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
@@ -229,7 +229,7 @@ export const emailData: TEmail[] = [
     sender: 'Lea Hartmann',
     subject: 'Website Maintenance',
     preview: 'Hey, We’ll be performing website maintenance this weekend. Please be aware of...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: true,
     isImportant: false,
@@ -240,7 +240,7 @@ export const emailData: TEmail[] = [
     sender: 'Felix Lange',
     subject: 'Customer Feedback',
     preview: 'Hi, We’ve received some positive customer feedback. Check out the attached...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: true,
@@ -251,7 +251,7 @@ export const emailData: TEmail[] = [
     sender: 'Emilia Wagner',
     subject: 'Office Closure Notice',
     preview: 'Hello, Please be informed that the office will be closed on [Date] for [Reason]...',
-    date: 'Sun Mar 23 2025 15:31:25 GMT+0600 (Bangladesh Standard Time)',
+    date: '2025-03-23T09:31:25.000Z',
     isRead: true,
     isStarred: false,
     isImportant: false,
